requirements.txt
praw
discord
python-dotenv